修改BUG：字典表数据未保护，限制为仅Administrators成员可删除字典数据

+++ Bootstrap.Admin/Controllers/DictsController.cs
@@ -1,5 +1,8 @@
 ﻿using Bootstrap.Admin.Models;
 using Bootstrap.DataAccess;
+using Longbow.Security.Principal;
+using Newtonsoft.Json.Linq;
+using System.Collections.Generic;
 using System.Linq;
 using System.Web.Http;
 
@@ -39,11 +42,32 @@ namespace Bootstrap.Admin.Controllers
         /// <summary>
         /// 
         /// </summary>
+        /// <returns></returns>
+        [HttpPost]
+        public IEnumerable<string> Post(int id, [FromBody]JObject value)
+        {
+            var ret = new List<string>();
+            dynamic json = value;
+            switch ((string)json.type)
+            {
+                case "category":
+                    ret = DictHelper.RetrieveCategories().ToList();
+                    break;
+                default:
+                    break;
+            }
+            return ret;
+        }
+        /// <summary>
+        /// 
+        /// </summary>
         /// <param name="id"></param>
         [HttpDelete]
-        public bool Delete([FromBody]string value)
+        public dynamic Delete([FromBody]string value)
         {
-            return DictHelper.DeleteDict(value);
+            if (!LgbPrincipal.IsAdmin(User.Identity.Name) && !User.IsInRole("Administrators")) return new { result = false, msg = "当前用户权限不够" };
+            var result = DictHelper.DeleteDict(value);
+            return new { result = result, msg = result ? "成功！" : "失败" };
         }
     }
 }
